requirements.txt
Django>=6.0,<7
djangorestframework>=3.15
djangorestframework-simplejwt>=5.3
